Move `/linkaccount`
